+++ spec/support/file_helper.rb
@@ -11,8 +11,6 @@ module FileHelper
 
     File.open(file_path, 'w') do |file|
       case content
-      when ''
-        # Write nothing. Create empty file.
       when String
         file.puts content
       when Array